+++ src/workers/blurhash.ts
@@ -1,0 +1,9 @@
+import { expose } from 'comlink'
+import { ValidationError, encode, decode, isBlurhashValid } from 'blurhash'
+
+expose({
+    ValidationError,
+    encode,
+    decode,
+    isBlurhashValid,
+})